+++ js/main.js
@@ -1,91 +1,1 @@
-const DESCRIPTION = [
-  'Раз, два, три',
-  'Blue Is the Colour',
-  'Зараза',
-  'Мяу',
-  'Пиу пау'
-];
-
-const MESSAGE = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAME = [
-  'Никита',
-  'Джон',
-  'Фелисити',
-  'Фрэнк',
-  'Юлиан Альфред Панкрац, виконт де Леттенхоф',
-];
-
-// Функция, которая генерирует случайное целое число
-
-const getRandomInt = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-// Функция, которая генерирует уникальное число в указанном диапазоне
-
-function getUniqueId(min, max) {
-  let num = min;
-
-  return () => {
-    if(num <= max) {
-      return num++;
-    }
-  };
-}
-
-// Функция, которая выбирает случайный элемент массива
-
-const getRandomArrayElement = (arr) => arr[getRandomInt(0, arr.length - 1)];
-
-
-const getPhotoId = getUniqueId(1, 25);
-const getCommentId = getUniqueId(1, Infinity);
-
-// Функция для описания объекта с комментарием
-
-// Количество комментариев к каждой фотографии — случайное число от 0 до 30. Все комментарии генерируются случайным образом.
-// У каждого комментария есть идентификатор — id — любое число. Идентификаторы не должны повторяться.
-// Поле avatar — это строка, значение которой формируется по правилу img/avatar-{{случайное число от 1 до 6}}.svg. Аватарки подготовлены в директории img.
-// Для формирования текста комментария — message — необходимо взять одно или два случайных предложения
-// Имена авторов также должны быть случайными. Набор имён для комментаторов составьте сами. Подставляйте случайное имя в поле name
-
-
-function createComment () {
-  return {
-    id: getCommentId,
-    avatar: `img/avatar-${getRandomInt(1, 6)}.svg`,
-    message: getRandomArrayElement(MESSAGE, getRandomInt(1, 2)),
-    name: getRandomArrayElement (NAME)
-  };
-}
-
-// Функция для описания фотографии
-
-// id, число — идентификатор опубликованной фотографии. Это число от 1 до 25. Идентификаторы не должны повторяться.
-// url, строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25. Адреса картинок не должны повторяться.
-// description, строка — описание фотографии. Описание придумайте самостоятельно.
-// likes, число — количество лайков, поставленных фотографии. Случайное число от 15 до 200.
-// comments, массив объектов — список комментариев, оставленных другими пользователями к этой фотографии.
-
-function photoDescribe () {
-  return {
-    id: getPhotoId,
-    url: `/photos/${getPhotoId}.jpg`,
-    description: getRandomArrayElement(DESCRIPTION),
-    likes: getRandomInt (15, 200),
-    comments: Array.from({length: getRandomInt(0, 30)}, createComment)
-  };
-}
-
-photoDescribe();
+import './util.js';
